// src/entities/onboarding/ui/OnboardingCategoriesList/types.ts
export enum MyHealthCategory {
  BodyMeasurements = 'BodyMeasurements',
  Goals = 'Goals',
  SportAndActivity = 'SportAndActivity',
  Training = 'Training',
  Lifestyle = 'Lifestyle',
  Food = 'Food',
  MentalHealth = 'MentalHealth',
  WomenHealth = 'WomenHealth',
  OrgansAndSystems = 'OrgansAndSystems',
}